web/requirements/tags.txt
django-taggit>=0.23.0
django-taggit-serializer>=0.1.7